requirements.txt
# Usage: pip install -r requirements.txt

ensure
numpy==1.19.5
pandas==1.3.4
matplotlib>=3.2.2
opencv-python>=4.1.1
opencv-python-headless
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
pytesseract==0.3.8
torch==1.10.1,!=1.12.0
torchaudio==0.10.1
torchvision==0.11.2,!=0.13.0
tqdm>=4.41.0
protobuf<4.21.3
Flask
flask-restx
googletrans==3.1.0a0
randominfo
python-box
seaborn

transformers==4.16.2
sentencepiece==0.1.95

ipython  # interactive notebook
psutil  # system utilization
thop  # FLOPs computation

-e .